scripts/ModalAnalysis/poutre2d_2.py (update)
