+++ scripts/ModalAnalysis/poutre2d_2.py
@@ -14,6 +14,20 @@ import finite_elements.loads
 import finite_elements.analysis
 import finite_elements.conditions
 
+def find_points_with_x(nodes, x):
+    indices_x = []
+    for i, node in enumerate(nodes):
+        if node.x == x:
+            indices_x.append(i)
+    return indices_x
+
+def find_points_with_y(nodes, y):
+    indices_y = []
+    for i, node in enumerate(nodes):
+        if node.y == y:
+            indices_y.append(i)
+    return indices_y
+
 # %% Mesh2D
 
 files_path = ['../InputFiles/2D/poutre_2d_0.8', '../InputFiles/2D/poutre_2d_0.5',
@@ -30,9 +44,9 @@ mesh = gmsh.define_triangular_element_mesh()
 
 # %% Finite Element Mesh2D
 
-# elasticity_modulus, poisson_ratio, thickness, mass_density = 30*1e6, 0.25, 1, 2.7
+elasticity_modulus, poisson_ratio, thickness, mass_density = 30*1e6, 0.25, 1, 2.7
 # elasticity_modulus, poisson_ratio, thickness, mass_density = 70*1e6, 0.33, 1, 2700 #aluminium
-elasticity_modulus, poisson_ratio, thickness, mass_density = 210*1e9, 0.25, 1, 7860 #acier
+# elasticity_modulus, poisson_ratio, thickness, mass_density = 210*1e9, 0.25, 1, 7860 #acier
 
 
 group_elements = []
@@ -50,7 +64,8 @@ mesh.nodes = gmsh.nodes[0]['all_nodes'] #Keep Gmsh order
 mesh.node_to_index = {mesh.nodes[i]: i for i in range(len(mesh.nodes))}
 
 
-# %% Analysis
+# %% Modal Analysis
+# %%% No Loads / No Conditions
 
 analysis = fe.analysis.FiniteElementAnalysis(mesh, [], [], [], [], [], [],
                                              plane_strain=False, plane_stress=True)
@@ -58,18 +73,76 @@ analysis = fe.analysis.FiniteElementAnalysis(mesh, [], [], [], [], [], [],
 eigvals, eigvecs = analysis.modal_analysis()
 elasticity_results = []
 
-for eigvec in eigvecs.T[0:10]:
+for eigvec in eigvecs[0:10]:
     elasticity_results.append(fe.results.ElasticityResults2D(analysis.mesh,
                                                              eigvec,
                                                              analysis.plane_strain, analysis.plane_stress))
 
 for elasticity_result in elasticity_results:
-    # elasticity_result.plot_deformed_mesh(amplitude=50)
-    elasticity_result.plot_displacement_per_node_xy(amplitude=50)
+    # elasticity_result.plot_deformed_mesh(amplitude=1)
+    elasticity_result.plot_displacement_per_node_xy(amplitude=1)
+
+# %%% With Conditions
+
+application_indices = find_points_with_x(nodes=mesh.nodes, x=0)
+application_indices.extend(find_points_with_x(nodes=mesh.nodes, x=10))
+
+node_boundary_conditions = []
+for index in application_indices:
+    node_boundary_conditions.extend([finite_elements.conditions.NodeBoundaryCondition(mesh.nodes[index], 0, 1),
+                                     finite_elements.conditions.NodeBoundaryCondition(mesh.nodes[index], 0, 2)])
+
+# load = -1000
+# application_index = mesh.node_to_index[volmdlr.mesh.Node2D(10,1)]
+# node_loads = [fe.loads.NodeLoad(mesh.nodes[application_index], load, 2)]
+
+analysis = fe.analysis.FiniteElementAnalysis(mesh, [], [], [], [], node_boundary_conditions, [],
+                                             plane_strain=False, plane_stress=True)
+
+eigvals, eigvecs = analysis.modal_analysis()
+elasticity_results = []
+
+for eigvec in eigvecs[0:10]:
+    elasticity_results.append(fe.results.ElasticityResults2D(analysis.mesh,
+                                                             eigvec,
+                                                             analysis.plane_strain, analysis.plane_stress))
+
+for elasticity_result in elasticity_results:
+    # elasticity_result.plot_deformed_mesh(amplitude=1)
+    elasticity_result.plot_displacement_per_node_xy(amplitude=1)
+
+# %%% With Conditions + Loads
+
+# application_indices = find_points_with_x(nodes=mesh.nodes, x=0)
+
+# node_boundary_conditions = []
+# for index in application_indices:
+#     node_boundary_conditions.extend([finite_elements.conditions.NodeBoundaryCondition(mesh.nodes[index], 0, 1),
+#                                      finite_elements.conditions.NodeBoundaryCondition(mesh.nodes[index], 0, 2)])
+
+# load = -1000
+# application_index = mesh.node_to_index[volmdlr.mesh.Node2D(10,1)]
+# node_loads = [fe.loads.NodeLoad(mesh.nodes[application_index], load, 2)]
+
+# analysis = fe.analysis.FiniteElementAnalysis(mesh, [], [], [], [], node_boundary_conditions, [],
+#                                              plane_strain=False, plane_stress=True)
+
+# eigvals, eigvecs = analysis.modal_analysis()
+# elasticity_results = []
+
+# for eigvec in eigvecs[0:10]:
+#     elasticity_results.append(fe.results.ElasticityResults2D(analysis.mesh,
+#                                                              eigvec,
+#                                                              analysis.plane_strain, analysis.plane_stress))
+
+# for elasticity_result in elasticity_results:
+#     # elasticity_result.plot_deformed_mesh(amplitude=1)
+#     elasticity_result.plot_displacement_per_node_xy(amplitude=1)
+
 
 # %% VTK files generation
 
-for i, elasticity_result in enumerate(elasticity_results):
-    elasticity_result.update_vtk_with_results(
-        input_file_name = file_path+'.vtk',
-        output_file_name = file_path+'_mode_n°_'+str(i)+'.vtk')
+# for i, elasticity_result in enumerate(elasticity_results):
+#     elasticity_result.update_vtk_with_results(
+#         input_file_name = file_path+'.vtk',
+#         output_file_name = file_path+'_mode_n°_'+str(i)+'.vtk')
